health-chatbot/requirements.txt
Flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
google-cloud-translate==3.12.0
googletrans==3.1.0a0 